refactor: Split up serve and InferenceServicer into different files for better unit testing later on.
